Fixed breaking issue & other minor issues
