+++ config.js
@@ -1,0 +1,1 @@
+export const Api_Key = "2dddca87eac46f4126ae969f4066037a";
